+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## (Unreleased) ##
+
+* Allows the user to specify ignored properties and selectors
+
 ## 0.2.1 - 3/28/2013 ##
 
 * Changes coloring to the selectors and declarations